+++ docs/source/getting-started/quickstart/template-root/src/main/java/com/digitalasset/quickstart/iou/IouMain.java
@@ -81,7 +81,7 @@ public class IouMain {
 
         Disposable ignore = client.getTransactionsClient().getTransactions(acsOffset.get(), iouFilter, true)
                 .forEach(t -> {
-                    for (Event event : t.getEvents()) {
+                    for (FlatEvent event : t.getEvents()) {
                         if (event instanceof CreatedEvent) {
                             CreatedEvent createdEvent = (CreatedEvent) event;
                             long id = idCounter.getAndIncrement();