+++ language-support/java/codegen/src/ledger-tests/scala/com/digitalasset/CodegenLedgerTest.scala
@@ -18,6 +18,7 @@ import com.daml.ledger.javaapi.data.{
   Event,
   Filter,
   FiltersByParty,
+  FlatEvent,
   GetTransactionsRequest,
   LedgerOffset,
   NoFilter,
@@ -111,8 +112,8 @@ class CodegenLedgerTest extends FlatSpec with Matchers {
             .fromProto(r)
             .getTransactions
             .stream())
-      .flatMap[Event]((t: Transaction) => t.getEvents.stream)
-      .flatMap[CreatedEvent]((e: Event) =>
+      .flatMap[FlatEvent]((t: Transaction) => t.getEvents.stream)
+      .flatMap[CreatedEvent]((e: FlatEvent) =>
         e match {
           case e: CreatedEvent => singletonList(e).stream();
           case _ => Collections.emptyList[CreatedEvent]().stream()